Cloned the repo and began edits
This morning I added to the README of the repo
The first thing I will beging working on is the landing page of my personal website.
I plan on documenting the process of making the lading page by bloggng about it.
The blog will be apart of my personal website
Next steps:
- Plan the out line of the website
- Begin planning for the landing page first